pyproject.toml
[project]
name = "project-creator-core"
version = "0.0.1"
description = "ADHD Framework core: project_creator_core"
requires-python = ">=3.11"
dependencies = [
    "logger-util",
    "creator-common-core",
    "exceptions-core",
    "github-api-core",
    "modules-controller-core",
    "pyyaml>=6.0",
]

[project.urls]
Repository = "https://github.com/AI-Driven-Highspeed-Development/project_creator_core.git"

[tool.adhd]
layer = "dev"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.build]
dev-mode-dirs = [".."]
